Keep aliases at the end of classes
We had a loose policy in the corelib to keep method definitions
alphabetized, that doesn't work well with aliases in which you should
alphabetize using the second argument, but only after the original
method is defined.

So with this commit we move all the aliases to the end of the class
definition, where all the original methods are already defined and
we keep them alphabetized.

Some aliases are skipped as they are used to enhance the original
method and need to happen before the new definition.